Move some code in its own file.
